feat: split util functions to individual files
